+++ api/src/opentrons/hardware_control/emulation/module_server/__init__.py
@@ -1,0 +1,8 @@
+from .server import ModuleStatusServer
+from .client import ModuleServerClient
+
+
+__all__ = [
+    "ModuleStatusServer",
+    "ModuleServerClient",
+]
